Fixing the ALLOCARGV opcode in all representations.
git-svn-id: svn+ssh://saulius-grazulis.lt/home/saulius/svn-repositories/compilers/trunk@4323 f7c8cbe2-f9e6-0310-9050-d7f47bf15973

+++ starta/interpreters/split-stackcell/tcodes.c
@@ -1820,10 +1820,10 @@ int JMP( INSTRUCTION_FN_ARGS )
 
 int ALLOCARGV( INSTRUCTION_FN_ARGS )
 {
-    ssize_t first = istate.ep[1].num.ssize;
-    ssize_t last = istate.ep[0].num.ssize;
+    int first = istate.ep[1].num.ssize;
+    int last = istate.ep[0].num.ssize;
     char **argv = NULL;
-    ssize_t i;
+    int i;
 
     TRACE_FUNCTION();
 
@@ -1833,14 +1833,14 @@ int ALLOCARGV( INSTRUCTION_FN_ARGS )
     if( last < 0 || last > istate.argc ) last = istate.argc;
 
     if( last >= first && istate.argv != NULL ) {
-        ssize_t length = last - first + 1;
+        int length = last - first + 1;
 	argv = bcalloc_array( REF_SIZE, length, 1, EXCEPTION );
 
 	BC_CHECK_PTR( argv );
 	STACKCELL_SET_ADDR( istate.ep[0], argv );
 
 	for( i = first; i <= last; i++ ) {
-            ssize_t k = i - first;
+            int k = i - first;
 	    argv[k] = bcalloc_blob( strlen(istate.argv[i]) + 1, EXCEPTION );
 	    BC_CHECK_PTR( argv[k] );
 	    strcpy( argv[k], istate.argv[i] );
